# backend/public/swagger.json
{
	"components": {
...
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"tokenVersion": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"messagesSent": {
						"items": {
							"$ref": "#/components/schemas/Message"
						},
						"type": "array"
					},
					"messagesReceived": {
						"items": {
							"$ref": "#/components/schemas/Message"
						},
						"type": "array"
					},
					"profilePictureId": {
						"type": "number",
						"format": "double"
					},
					"profilePicture": {
						"$ref": "#/components/schemas/Picture"
					}
				},
				"required": [
					"id",
					"email",
					"password",
					"roleId",
					"role",
					"tokenVersion",
					"messagesSent",
					"messagesReceived"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"role",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"senderId": {
						"type": "number",
						"format": "double"
					},
					"receiverId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"delivered": {
						"type": "boolean"
					},
					"read": {
						"type": "boolean"
					},
					"sender": {
						"$ref": "#/components/schemas/User"
					},
					"receiver": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"content",
					"senderId",
					"receiverId",
					"createdAt",
					"delivered",
					"read",
					"sender",
					"receiver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Listing": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"statusId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/ListingStatus"
					},
					"personId": {
						"type": "number",
						"format": "double"
					},
					"person": {
						"$ref": "#/components/schemas/Person"
					},
					"pictures": {
						"items": {
							"$ref": "#/components/schemas/Picture"
						},
						"type": "array"
					},
					"faultCategoryId": {
						"type": "number",
						"format": "double"
					},
					"faultCategory": {
						"$ref": "#/components/schemas/FaultCategory"
					},
					"offers": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					},
					"cityId": {
						"type": "number",
						"format": "double"
					},
					"city": {
						"$ref": "#/components/schemas/City"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"createdAt",
					"updatedAt",
					"statusId",
					"status",
					"personId",
					"person",
					"pictures",
					"faultCategoryId",
					"faultCategory",
					"offers",
					"cityId",
					"city"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListingStatus": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"status",
					"listings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Person": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/Review"
						},
						"type": "array"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"userId",
					"user",
					"createdAt",
					"updatedAt",
					"reviews",
					"listings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Service": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"oib": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"cityId": {
						"type": "number",
						"format": "double"
					},
					"city": {
						"$ref": "#/components/schemas/City"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/Review"
						},
						"type": "array"
					},
					"faultCategories": {
						"items": {
							"$ref": "#/components/schemas/FaultCategory"
						},
						"type": "array"
					},
					"offers": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"oib",
					"phone",
					"address",
					"cityId",
					"city",
					"userId",
					"user",
					"createdAt",
					"updatedAt",
					"reviews",
					"offers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"City": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"postalCode",
					"services",
					"listings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Review": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"authorId": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"$ref": "#/components/schemas/Person"
					},
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					}
				},
				"required": [
					"id",
					"content",
					"score",
					"createdAt",
					"updatedAt",
					"authorId",
					"author",
					"serviceId",
					"service"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FaultCategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"parent": {
						"$ref": "#/components/schemas/FaultCategory"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/FaultCategory"
						},
						"type": "array"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Offer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"statusId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/OfferStatus"
					},
					"listingId": {
						"type": "number",
						"format": "double"
					},
					"listing": {
						"$ref": "#/components/schemas/Listing"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"price",
					"serviceId",
					"service",
					"createdAt",
					"updatedAt",
					"statusId",
					"status",
					"listingId",
					"listing"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OfferStatus": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"offers": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"status",
					"offers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Picture": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"cloudinaryId": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"listingId": {
						"type": "number",
						"format": "double"
					},
					"listing": {
						"$ref": "#/components/schemas/Listing"
					}
				},
				"required": [
					"id",
					"name",
					"cloudinaryId",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IListingPayload": {
				"properties": {
					"personId": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"faultCategoryId": {
						"type": "number",
						"format": "double"
					},
					"cityId": {
						"type": "number",
						"format": "double"
					},
					"statusId": {
						"type": "number",
						"format": "double"
					},
					"pictures": {
						"items": {
							"$ref": "#/components/schemas/Picture"
						},
						"type": "array"
					}
				},
				"required": [
					"personId",
					"title",
					"description",
					"faultCategoryId",
					"cityId",
					"statusId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IListingPaginatedResult": {
				"properties": {
					"current_page": {
						"type": "number",
						"format": "double"
					},
					"total_pages": {
						"type": "number",
						"format": "double"
					},
					"per_page": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"current_page",
					"total_pages",
					"per_page",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IMessagePayload": {
				"properties": {
					"content": {
						"type": "string"
					},
					"senderId": {
						"type": "number",
						"format": "double"
					},
					"receiverId": {
						"type": "number",
						"format": "double"
					},
					"delivered": {
						"type": "boolean"
					},
					"read": {
						"type": "boolean"
					}
				},
				"required": [
					"content",
					"senderId",
					"receiverId",
					"delivered",
					"read"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/pictures/listing/{id}": {
			"get": {
				"operationId": "GetPicturesByListing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Picture"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Picture"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/services/search": {
			"get": {
				"operationId": "GetSearchResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "service",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "faultCategoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cityId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/listings": {
			"get": {
				"operationId": "GetListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Listing"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateListing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Listing"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IListingPayload"
							}
						}
					}
				}
			}
		},
		"/listings/active/{id}": {
			"get": {
				"operationId": "GetActiveListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IListingPaginatedResult"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/history/{id}": {
			"get": {
				"operationId": "GetHistoryListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IListingPaginatedResult"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listings/id/{id}": {
			"get": {
				"operationId": "GetListing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Listing"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/listings/upload-pictures/{id}": {
			"post": {
				"operationId": "UploadPictures",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Picture"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/listings/search": {
			"get": {
				"operationId": "GetSearchResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IListingPaginatedResult"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "listing",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "faultCategoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cityId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/messages": {
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Message"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Message"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					}
				},
				"tags": [
					"Message"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IMessagePayload"
							}
						}
					}
				}
			}
		},
		"/messages/id/{id}": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Message"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Message"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/messages/user/{id}": {
			"get": {
				"operationId": "GetUserMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Message"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/messages/contacts": {
			"get": {
				"operationId": "GetContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Message"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/messages/contacts/{id}": {
			"get": {
				"operationId": "GetContactById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Message"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}